better livestream support, fullscreen and zoom buttons
