+++ src/app/auth/AuthHeader.tsx
@@ -1,4 +1,4 @@
-import { PropsWithChildren } from "react";
+import { type PropsWithChildren } from "react";
 import { CardHeader } from "~/components/ui/card";
 
 export function AuthHeader({ children }: PropsWithChildren) {
