--- src/components/youtube.ts
@@ -1,7 +1,0 @@
-export const YOUTUBE_PLAYER_STATES = {
-  ENDED: 0,
-  PLAYING: 1,
-  PAUSED: 2,
-  BUFFERING: 3,
-  CUED: 5,
-} as const;
