+++ src/lib/transformSecondsToFormattedTime.ts
@@ -1,13 +1,15 @@
 export function transformSecondsToFormattedTime(seconds: number): string {
-  const hours = Math.floor(seconds / 3600);
+  const days = Math.floor(seconds / (3600 * 24));
+  const hours = Math.floor((seconds % (3600 * 24)) / 3600);
   const minutes = Math.floor((seconds % 3600) / 60);
   const remainingSeconds = seconds % 60;
 
+  const formattedDays = days > 0 ? `${days.toString().padStart(2, "0")}:` : "";
   const formattedHours =
     hours > 0 ? `${hours.toString().padStart(2, "0")}:` : "";
   const formattedMinutes = `${minutes.toString().padStart(2, "0")}:`;
   let formattedSeconds = remainingSeconds.toString().split(".")?.[0] ?? "";
   formattedSeconds = formattedSeconds.padStart(2, "0");
 
-  return `${formattedHours}${formattedMinutes}${formattedSeconds}`;
+  return `${formattedDays}${formattedHours}${formattedMinutes}${formattedSeconds}`;
 }
